pa2 finish

+++ abstract-machine/am/src/platform/nemu/ioe/timer.c
@@ -15,8 +15,8 @@ typedef struct {
 */
 void __am_timer_uptime(AM_TIMER_UPTIME_T *uptime) {
   // guarantee visit high first, that is when time is updated
-  uint64_t const _read_time_hi = *(uint32_t *)(RTC_ADDR + 4);
-  uint64_t const _read_time_lo = *(uint32_t *)(RTC_ADDR);
+  uint64_t const _read_time_hi = *(uint32_t *)(uintptr_t)(RTC_ADDR + 4);
+  uint64_t const _read_time_lo = *(uint32_t *)(uintptr_t)(RTC_ADDR);
 
   uint64_t const _read_time = (_read_time_hi << 32) | _read_time_lo;
   uptime->us = _read_time ;//- boot_time.us;
